refactor(web-analytics): move bot definitions and fixtures out of hogql/ to avoid hogql review
