+++ posthog/hogql/functions/test/test_traffic_type_real_ua.py
@@ -4,13 +4,14 @@ from uuid import uuid4
 import pytest
 from posthog.test.base import BaseTest, _create_event, flush_persons_and_events
 
-from posthog.hogql.functions.bot_ua_fixtures import (
+from posthog.hogql.query import execute_hogql_query
+
+from posthog.hogql_queries.web_analytics.bot_definitions import BOT_DEFINITIONS
+from posthog.hogql_queries.web_analytics.bot_ua_fixtures import (
     BOT_USER_AGENTS,
     CATEGORY_TO_TRAFFIC_CATEGORY,
     CATEGORY_TO_TRAFFIC_TYPE,
 )
-from posthog.hogql.functions.traffic_type import BOT_DEFINITIONS
-from posthog.hogql.query import execute_hogql_query
 
 
 def _find_matching_pattern(ua: str) -> str | None:
